# study-service/src/main/resources/openapi/study-api.openapi.yml
openapi: 3.0.0
info:
  title: Study API
  version: 0.0.1
servers:
  - url: http://localhost:8083/api/study
paths:
  /schedule:
    get:
      security:
        - bearerAuth: [ ]
      summary: Get All Schedule
      tags:
        - Schedule
      operationId: getAllSchedule
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleOpenAPI"
        '400':
          description: Bad Request
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
    post:
      security:
        - bearerAuth: [ ]
      summary: Create a Schedule
      tags:
        - Schedule
      operationId: createSchedule
      requestBody:
        description: Schedule Object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleOpenAPI'
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
    put:
      security:
        - bearerAuth: [ ]
      summary: Update a Schedule by Id
      tags:
        - Schedule
      operationId: updateSchedule
      requestBody:
        description: Schedule Object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleOpenAPI'
      responses:
        '200':
          description: Updated
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
  /schedule/{id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Schedule Information by Id
      tags:
        - Schedule
      operationId: getScheduleByGroupId
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
    delete:
      security:
        - bearerAuth: [ ]
      summary: Delete a Schedule by Id
      tags:
        - Schedule
      operationId: deleteSchedule
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: Deleted
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
  /schedule/professor/{professorId}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Get Schedule By Professor Id
      tags:
        - Schedule
      operationId: getScheduleByProfessorId
      parameters:
        - in: path
          name: professorId
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleOpenAPI"
        '400':
          description: Bad Request
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
...
  /attendance-statistics/schedule/{scheduleId}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Get Attendance Statistics by Schedule Id
      tags:
        - AttendanceStatistics
      operationId: getStatisticsBySchedule
      parameters:
        - in: path
          name: scheduleId
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceStatisticsOpenApi"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
...
  /materials/{id}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Materials Information
      tags:
        - Materials
      operationId: getMaterialsById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyMaterialsOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found

    delete:
      security:
        - bearerAuth: [ ]
      summary: Delete a Material by Id
      tags:
        - Materials
      operationId: deleteMaterials
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: Deleted
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
  /homework/materials/{materialsId}/all:
    get:
      security:
        - bearerAuth: [ ]
      summary: All Homeworks Information by Materials Id
      tags:
        - Homework
      operationId: getAllHomeworksByMaterials
      parameters:
        - in: path
          name: materialsId
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HomeworkOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
  /homework/materials/{materialsId}:
    post:
      security:
        - bearerAuth: [ ]
      summary: Create a Homework
      tags:
        - Homework
      operationId: createHomework
      parameters:
        - in:
            path
          name: materialsId
          schema:
            type: integer
            format: int64
          required: true
      requestBody:
        description: Homework Object
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SaveHomeworkRequestOpenApi'
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
  /homework/grade/{homeworkId}:
    put:
      security:
        - bearerAuth: [ ]
      summary: Grade a Homework
      tags:
        - Homework
      operationId: gradeHomework
      parameters:
        - in: path
          name: homeworkId
          schema:
            type: integer
            format: int64
          required: true
      requestBody:
        description: Grade and comment
        required: true
        content:
          application/json:
            schema:
              $ref:  '#/components/schemas/GradeHomeworkRequestOpenApi'
      responses:
        '200':
          description: Updated
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
  /homework/{homeworkId}:
    put:
      security:
        - bearerAuth: [ ]
      summary: Update a Homework
      tags:
        - Homework
      operationId: updateHomework
      parameters:
        - in: path
          name: homeworkId
          schema:
            type: integer
            format: int64
          required: true
      requestBody:
        description: Homework Object
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SaveHomeworkRequestOpenApi'
      responses:
        '200':
          description: Updated
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
    get:
      security:
        - bearerAuth: [ ]
      summary: Homework Information
      tags:
        - Homework
      operationId: getHomeworkById
      parameters:

        - in: path
          name: homeworkId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeworkOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
    delete:
      security:
        - bearerAuth: [ ]
      summary: Delete a Homework
      tags:
        - Homework
      operationId: deleteHomework
      parameters:
        - in: path
          name: homeworkId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Deleted
        '400':
          description: Bad Request
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
  /homework/student/{studentId}:
    get:
      security:
        - bearerAuth: [ ]
      summary: All Homeworks Information By Student
      tags:
        - Homework
      operationId: getHomeworkByStudentId
      parameters:
        - name: studentId
          in: path
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HomeworkOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Student Not Found
  /homework/materials/{materialsId}/student/{studentId}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Get Student's Homework by Material
      tags:
        - Homework
      operationId: getHomeworkByStudentAndMaterials
      parameters:
          - name: materialsId
            in: path
            schema:
              type: integer
              format: int64
            required: true
          - name: studentId
            in: path
            schema:
              type: integer
              format: int64
            required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HomeworkOpenAPI"
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Student Not Found
...
  /download/homework/{fileId}:
    get:
      security:
        - bearerAuth: [ ]
      summary: Download File by It's Id
      tags:
        - File Download
      operationId: downloadHomeworkFileById
      parameters:
        - in: path
          name: fileId
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
                format: binary
        '401':
          $ref: "#/components/responses/UnauthorizedError"
        '403':
          description: Do Not Have Necessary Rights to Execute This Operation
        '404':
          description: Not Found
...
components:
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ScheduleOpenAPI:
      type: object
      required:
        - dateTime
        - studyGroupId
        - subjectId
        - classFormat
        - professorId
      properties:
        id:
          type: integer
          format: int64
        date:
          type: string
          format: date-time
        studyGroupId:
          type: integer
          format: int64
        subjectId:
          type: integer
          format: int64
        classFormat:
          $ref: "#/components/schemas/ClassFormat"
        auditorium:
          type: string
        linkForTheClass:
          type: string
        professorId:
          type: integer
          format: int64
...
    ClassFormat:
      type: string
      enum:
        - IN_PERSON
        - DISTANT
    StudyMaterialsType:
      type: string
      enum:
        - MATERIAL
        - TASK
    StudyMaterialsOpenAPI:
      type: object
      required:
        - type
        - subjectId
        - materialsText
      properties:
        id:
          type: integer
          format: int64
        type:
          $ref: "#/components/schemas/StudyMaterialsType"
        materialsText:
          type: string
        subjectId:
          type: integer
          format: int64
        homeworkIds:
          type: array
          items:
            type: integer
            format: int64
        filesId:
          type: array
          items:
            type: integer
            format: int64
        filesNames:
          type: array
          items:
            type: string
    HomeworkOpenAPI:
      type: object
      required:
        - taskId
        - studentId
        - date
      properties:
        id:
          type: integer
          format: int64
        taskId:
          type: integer
          format: int64
        studentId:
          type: integer
          format: int64
        date:
          type: string
          format: date-time
        grade:
          type: integer
        comment:
          type: string
        filesNames:
          type: array
          items:
            type: string
        fileIds:
          type: array
          items:
            type: integer
            format: int64
    SaveHomeworkRequestOpenApi:
      type: object
      required:
        - homework
        - file
      properties:
        homework:
          $ref: "#/components/schemas/HomeworkOpenAPI"
        file:
          type: array
          items:
            type: string
            format: binary
            description: File with homework
...
    AttendanceStatisticsOpenApi:
      type: object
      required:
        - scheduleId
        - userId
        - status
      properties:
        scheduleId:
           type: integer
           format: int64
        status:
          $ref: "#/components/schemas/AttendanceStatusOpenApi"
        studentId:
          type: integer
          format: int64
    AttendanceStatusOpenApi:
      type: string
      enum:
        - ATTENDED
        - MISSED
    GradeHomeworkRequestOpenApi:
      type: object
      required:
        - grade
      properties:
        grade:
          type: integer
        comment:
          type: string